pyproject.toml
[tool.poetry]
name = "greek_stemmer_plus"
version = "0.2.0"
description = "Python Greek stemmer."
authors = ["German Attanasio <germanattanasio@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/germanattanasio/python_greek_stemmer"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
]
keywords = ["stemmer", "greek", "nlp"]

include = [
  "README.md",
  "greek_stemmer_plus/**/*.py",
  "greek_stemmer_plus/**/*.yml",
]


[tool.poetry.dependencies]
python = ">3.8"
PyYAML = ">5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
